+++ CHANGELOG.rst
@@ -3,6 +3,12 @@ Changelog for Metview's Python interface
 
 ========================================
 
+1.16.0 (unreleased)
+-------------------
+- added support for reading PNG files
+- added string() function from Metview
+
+
 1.15.0
 ------------------
 - Added arguments() function to mimic the behaviour of the Macro function of the same name